+++ src/option_pricing/models/heston/__init__.py
@@ -3,11 +3,11 @@ yap yap re-export core funcs
 """
 
 from .charfunc import HestonCharFn
-from .fourier import P_j
+from .fourier import P_j_Scalar
 from .params import HestonParams
 
 __all__ = [
-    "P_j",
+    "P_j_Scalar",
     "HestonParams",
     "HestonCharFn",
 ]